The powerful yet somehow effeminate fingers of the budding software developer graced his keyboard with all the dexterity and grace of a concert pianist.
Just minutes earlier, he had been feeling overwhelmed. The task he had just
been handed was to build a game in JavaScript — something he had never done
before, in a language he barely knew.

Yet now, as inspiration began to strike, the look of concern that had been
encroaching on his face was fighting a losing battle with a firm expression
of steely determination. He had had an idea.

The aformentioned nimble digits were in full swing as the right index
carved out a trajectory once unfamiliar yet now seared into its muscle memory
as it deftly churned out a forwards slash, follow in quick succession by an
asterisk and ten hyphens. One cheeky space later, the baton of this typing
relay was passed the immediate neighbour of Right Index, the infamous Middle
Finger, who in characteristically subversive style mimicked the exact
sequence of keystroke just type, but in reverse. It was fortunate for Right
Hand that this was exactly what was required at this moment, and as Middle
Finger landed firmly on Forwards Slash, the text that has ertswhile appeared
on the screen underwent a sea-change of transformations. The only visible
indicator that anything had happened was in the form of a visual cue in the
text turning to an overcast, understated greyish-blue colour.

All of this was happening automatically for the young, girly-handed coder.
His mind was preoccupied, roaming in an altogether different territory.

For he had had an idea for a game, and one which he was now excited to work
on. In his minds eye, he pictured a screen with words slowly fading into view
, some strange and unfamiliar, some altogether unreal. It was hard to tell
which words were real, and which were fake.

Therein lay the proverbial rub. The aim of this game, which was yet to have
even a single keystroke of its JavaScripty logic encoded, was to separate
reality from fiction, in a world where the two had come to overlap to a
disconcerting extent. What could be more insidious than fake news?

How about fake words?

---Do such things exist?--- I hear you cry. They do. A neural network trained
on a dictionary can now spin out entirely artificial words with invented
meanings that are utterly convincing. Fake News is old school- this is fake
language, created by a machine.

In an era where literacy is declining, this game could be the last bastion
of defence against the forces that are dumbing down society. In a world
where media outlets are bringing their language standards down to
make them more accessible, rather than acting as shining examples of classic
prose, this game could be the moment of change, the singularity which turns
a generation helplessly dependent on deceptively benign-looking commentary,
spoon-fed to them by the dubiously unbiased BBC, into masters of words,
with tongues as sharp as knives and minds ready to combat whatever the status
quo-led media would have us unthinkingly accept.

No more will we be dictated to by the evil i.

No more do we need meaningless, pandering aphorisms as a Guardian.

We can fight for ourselves. We can overcome. We can defend ourselves against

THE AI INVASION.

(c) David Wildman 2020. All rights carelessly overlooked.