These messages are getting unmanageably long. I cannot keep up with what I have started.
It was going to be so epic. The story of how I coded my first JavaScript game,
told in a series of excessively eloquent commit messages. But I can't
keep this up. I'm spending so much time writing the story that I am rapidly running
out of time to actually code the game...

...thought David to himself as he began writing his next commit message.

But he was a sucker for a blank canvas, and as the words spilled onto
the console, he couldn't help but continue the telling of the story
he had only seconds earlier sworn off.

David was fairly please with himself, all things considered. He had refactored
his objects, formed from two classes which were extended from a base class,
into an array of objects which was now being imported from a separate file.

I am not doing this to impress anyone

'Goodness' he thought to himself, 'I need to stop pressing the
up and down arrows. They mess everything up'Someone should make a name for this sensation', he thought to himself, blissfully unaware of the wall which he was rapidly approaching.

I decided it was time to end this facade and admit I had no idea what was going on