This seems to be going well!
I have the core function of the game sorted out.

When my function is called, I get a randomly-sorted array of objects.

Each object has properties including a word, its definition and x/y coordinates.

The next piece of the puzzle is to add a method to the global class for these objects
which will return a useful string of html which I can use to place it onto
the page, positioned absolutely (relative to its parent) at the given coordinates.

My fluffy penguin is giving me a hard stare.

I should probably go to bed before too long.